move code out of player character
