+++ src/components/character-builder/character-fundamentals/ancestry-choice-modal.tsx
@@ -1,17 +1,14 @@
 import { Ancestry } from '@/models/db/ancestry'
-import { PlayerCharacter } from '@/models/player-character'
 import { useContext, useEffect, useState } from 'react'
 import { PlayerCharacterContext } from '../../character-display/player-character-context'
 import { FeatureChoiceModal } from '../feature-choice-modal'
+import { SourcedFeature } from '@/models/player-character'
+import { CharacterEntity } from '@/models/db/character-entity'
 
 export function AncestryChoiceModal({
   onUpdate,
 }: {
-  onUpdate: (
-    updateFunction: (
-      playerCharacter: PlayerCharacter
-    ) => Promise<PlayerCharacter>
-  ) => void
+  onUpdate: (updateFunction: (cloned: CharacterEntity) => void) => void
 }) {
   const { playerCharacter } = useContext(PlayerCharacterContext)
   const [ancestries, setAncestries] = useState<Ancestry[]>([])
@@ -33,9 +30,17 @@ export function AncestryChoiceModal({
         entities={ancestries}
         initialId={playerCharacter.getAncestryId()}
         onSave={(ancestry) =>
-          onUpdate((playerCharacter) =>
-            playerCharacter.updateAncestry(ancestry._id.toString())
-          )
+          onUpdate((character: CharacterEntity) => {
+            character.ancestry_id = ancestry._id.toString()
+            character.attributes.ancestry = []
+            character.languages = []
+            character.heritage_id = ''
+            character.features.filter(
+              (val: SourcedFeature) =>
+                val.source === 'ANCESTRY' &&
+                val.feature.type === 'ANCESTRY_FEAT_SELECTION'
+            )[0].feature.value = null
+          })
         }
       ></FeatureChoiceModal>
     </>
