+++ src/components/character-builder/character-fundamentals/background-choice-modal.tsx
@@ -1,23 +1,17 @@
 import { Background } from '@/models/db/background'
+import { CharacterEntity } from '@/models/db/character-entity'
 import { Feat } from '@/models/db/feat'
-import { PlayerCharacter, SourcedFeature } from '@/models/player-character'
+import { SourcedFeature } from '@/models/player-character'
 import { retrieveEntity } from '@/utils/services/reference-lookup.service'
 import { cloneDeep } from 'lodash'
 import { useContext, useEffect, useState } from 'react'
 import { PlayerCharacterContext } from '../../character-display/player-character-context'
 import { FeatureChoiceModal } from '../feature-choice-modal'
 import { FeatSubChoiceModal } from '../level/feat-subchoice-modal'
-import { CharacterEntity } from '@/models/db/character-entity'
 
 export function BackgroundChoiceModal({
-  onAsyncUpdate,
   onUpdate,
 }: {
-  onAsyncUpdate: (
-    updateFunction: (
-      playerCharacter: PlayerCharacter
-    ) => Promise<PlayerCharacter>
-  ) => void
   onUpdate: (updateFunction: (cloned: CharacterEntity) => void) => void
 }) {
   const { playerCharacter } = useContext(PlayerCharacterContext)
@@ -56,9 +50,33 @@ export function BackgroundChoiceModal({
   }, [])
 
   const updateBackground = (background: Background) => {
-    onAsyncUpdate((playerCharacter: PlayerCharacter) =>
-      playerCharacter.updateBackground(background)
-    )
+    onUpdate((character: CharacterEntity) => {
+      character.background_id = background._id.toString()
+      character.attributes.background = []
+      character.features = character.features.filter(
+        (value) => value.source !== 'BACKGROUND'
+      )
+      character.features.push({
+        source: 'BACKGROUND',
+        feature: {
+          type: 'FEAT',
+          value: background.feat,
+          context: [],
+        },
+      })
+      character.features.push(
+        ...background.skills.map((value: ProficiencyFeatureValue) => {
+          return {
+            source: 'BACKGROUND',
+            feature: {
+              type: 'PROFICIENCY' as FeatureType,
+              value: value,
+            },
+          }
+        })
+      )
+    })
+
     reloadFeat(background.feat)
   }
 
@@ -92,7 +110,7 @@ export function BackgroundChoiceModal({
             onChange={handleSubChoiceChange}
           ></FeatSubChoiceModal>
         </div>
-      )}{' '}
+      )}
     </>
   )
 }
