+++ src/components/character-builder/level/level-section.tsx
@@ -21,52 +21,53 @@ export function LevelSection({
   )
   const { level } = useContext(CharacterLevelContext)
 
-  const loadCharacter = (updated: CharacterEntity) => {
-    const load: Promise<void> = (async () => {
-      updatePlayerCharacter(await PlayerCharacter.build(updated))
-    })()
-    wrapCharacterUpdate(load)
-  }
-
-  const handleFeatureUpdateForLevel =
-    (level?: number) =>
-    (matchingFunction: (source: SourcedFeature) => boolean) =>
-    (features: SourcedFeature[]) => {
-      let updated = cloneDeep(playerCharacter.getCharacter())
-
-      let toReplace = updated.features.filter((sourced) =>
-        matchingFunction(sourced)
-      )
-
-      if (level) {
-        toReplace = toReplace.filter(
-          (sourced) => sourced.feature.level == level
-        )
-      }
-
-      toReplace.forEach((item, idx) => {
-        const index = updated.features.indexOf(item)
-        updated.features.splice(index, 1, features[idx])
-      })
-      loadCharacter(updated)
-    }
-
-  const alternateHandleFeatureUpdate =
+  const handleFeatureUpdate =
     (original: SourcedFeature) => (replacement: SourcedFeature) => {
       let updated = cloneDeep(playerCharacter.getCharacter())
 
       const index = playerCharacter.getCharacter().features.indexOf(original)
       updated.features.splice(index, 1, replacement)
 
-      loadCharacter(updated)
+      const load: Promise<void> = (async () => {
+        updatePlayerCharacter(await PlayerCharacter.build(updated))
+      })()
+      wrapCharacterUpdate(load)
     }
 
   const handleSubclassChange = (subclass: Subclass) => {
-    const load: Promise<void> = (async () => {
-      const value: PlayerCharacter = await playerCharacter.updateSubclass(
-        subclass
+    const updated = cloneDeep(playerCharacter.getCharacter())
+    const subclassChoice = updated.features.find(
+      (value) => value.source === 'CLASS' && value.feature.type === 'SUBCLASS'
+    )
+    if (subclassChoice) {
+      subclassChoice.feature.value = subclass._id
+
+      updated.features = updated.features.filter(
+        (value) => value.feature.type !== 'SUBCLASS_FEATURE'
       )
-      updatePlayerCharacter(value)
+
+      const newFeatures = playerCharacter
+        .getClassEntity()
+        .features.filter((val) => val.type === 'SUBCLASS_FEATURE')
+
+      newFeatures.forEach((subclassFeature: Feature) => {
+        const matched = subclass.features.find(
+          (feature) => feature.name === subclassFeature.name
+        )
+        if (matched) {
+          subclassFeature.value = matched
+        }
+      })
+
+      updated.features.push(
+        ...newFeatures.map((feature: Feature) => {
+          return { source: 'CLASS', feature: feature }
+        })
+      )
+    }
+
+    const load: Promise<void> = (async () => {
+      updatePlayerCharacter(await PlayerCharacter.build(updated))
     })()
     wrapCharacterUpdate(load)
   }
@@ -90,7 +91,7 @@ export function LevelSection({
               <div>
                 <SubclassChoice
                   onSubclassChange={handleSubclassChange}
-                  onFeatureUpdate={alternateHandleFeatureUpdate}
+                  onFeatureUpdate={handleFeatureUpdate}
                 ></SubclassChoice>
               </div>
             ))}
@@ -106,7 +107,7 @@ export function LevelSection({
                   key={`skills-${index}`}
                   skillFeature={sourced.feature}
                   onSkillsUpdate={(feature: Feature) => {
-                    alternateHandleFeatureUpdate(sourced)({
+                    handleFeatureUpdate(sourced)({
                       source: 'CLASS',
                       feature: feature,
                     })
@@ -127,7 +128,7 @@ export function LevelSection({
                 {buildFeatChoice(
                   playerCharacter,
                   val,
-                  alternateHandleFeatureUpdate(val)
+                  handleFeatureUpdate(val)
                 )}
               </div>
             ))}
